[FEATURE] Add support to import namespaces for sniffs
During the development of sniffs for the CMS TYPO3 and the
framework FLOW3 we create some own codesniffs. Out folder
structure looks like:

|-TYPO3
|--Sniffs
|-TYPO3v4
|--ruleset.xml
|-FLOW3
|--ruleset.xml

Folder TYPO3 contains all sniffs. TYPO3v4 and FLOW3 implement
only a subset of the sniffs, which are located in FLOW3.
At the moment we have to symlink the TYPO3-Folder into the
PHP_CodeSniffer standard directory to use this. This
is not a nice solution.

This commit try to fix this problem. This commit introduces
to import sniff namespaces. Imagine you add your rules like
<rule ref="TYPO3.Files.OneClassPerFile" />

If there is no symlink in your standard directory for TYPO3
an exception will be thrown. With this commit you have the
possibility to import the namespace TYPO3 like
<namespace-import name="TYPO3" ref="../TYPO3/" />

Now the TYPO3 standard is known without symlinking :)