The generator don't work with relative paths in the ruleset.xml
If you call the codesniffer with the --generator option the getStandardFiles() function compiles a list of the sniffs via $phpcs->getSniffFiles($standardDir, $standard). As you can see it passes the standard dir and the name of the standard.

But the passed standard folder won't be used inside getSniffFiles(). At the end self::$standardDir in _expandRulesetReference() is empty and can't resolve sniffs which were specified relative in ruleset.xml.