Implementation of @codingStandardsIgnoreLine
If this suppression is used, only the following line will be suppressed, without the need for a closing *@codingStandardsIgnoreEnd* line.  This allows the suppression to occupy a little less space in the code, but was specifically implemented as a result of issue squizlabs/PHP_CodeSniffer#311.  Use of this suppression may also discourage large blocks of code being suppressed, or accidental suppression to the end of the file in the case where an *@codingStandardsIgnoreEnd* is misspelt or forgotten.

Commit includes updated unit tests.  The first attempt at this resulted in *@codingStandardsIgnoreLine* breaking any *@codingStandardsIgnoreStart*/*@codingStandardsIgnoreEnd* suppressions it was nested within, so there's a specific test for this case, too.

Alternative names for the suppression include *@codingStandardsIgnore* and *@codingStandardsIgnoreNext*.