Changing the regex check in the FIXME and TODO sniffs.
In fact, the problem is the use of the preg_match expression. As defined in the official PHP documentation :

> preg_match() returns 1 if the pattern matches given subject, 0 if it does not, or FALSE if an error occurred.

However the condition used in the sniffs is :

```
 if (preg_match('/(?:\A|[^\p{L}]+)fixme([^\p{L}]+(.*)|\Z)/ui', $content, $matches) !== 0) {
```

Which will in case of FALSE, enter the block code. Which it should not. I will submit a new PR correcting this behaviour.