Added experimental caching system between runs for better performance
On subsequent runs with the same config settings, a cache file can be used to only check files that have been modified.

Use the --cache command line argument to enable caching, or set the config var "cache" to have it on by default. If set in this way, use --no-cache to disable for a single run.