Allow indentation to be enforced over PHP blocks
Right now all the T_OPEN/T_CLOSE tokens lead, unconditionally,
to the indentation being reset to their level. In general,
it works ok, but there are standards that could prefer to
enforce multiple blocks indentation a bit harder.

There are situations where it may be desired to continue
checking indentation in a PHP block at the same level
the previous PHP block ended with. A classical example is
when you evaluate a condition in PHP, end the block, produce some
conditional output/html, just to continue in the next PHP block,
closing the condition scope or whatever.

This (opt-in) new setting (observePreviousPhpBlockIndentation) tries
to enforce that by looking for the indentation in the leaving block
when there are open scopes and applying it as default indentation in
the next incoming block. When there are not open scopes, there is no
restriction.

Surely some tests will show it better (next commit).