Tests for observePreviousPhpBlockIndentation
New fixture with the observePreviousPhpBlockIndentation setting enabled
(note I had to disable explicitly exact matching because it's apparently
dragged from previous test) that shows and covers its use when there
are remaining scopes opened.

Also it includes a fraction of ScopeIndentUnitTest.1.inc fixture
to ensure it does not break normal operations. I was not sure about
the best way to cover that.