Add Composer standards installer plugin.
This turns the Composer package into a plugin that comes with a custom installer that can install packages of type `phpcs-standard`.

A package that is of type `phpcs-standard` will need to have an `extra` key `standards` that contains an array of entries in the following format:
```JSON
	"extra"      : {
		"standards": {
			"<standard name 1>"      : "<standard path 1>",
			"<standard name 2>"      : "<standard path 2>"
		}
	}
```

The `<standard path>` is relative to the package root of the standard package.

For each of the standards defined in this way, the installer will create a symlink in `CodeSniffer/Standards/<standard name>` that points back to the corresponding standard.