Stopped gzipping files in the phar and just stripping whitespace instead.
The use of gzip compression was stopping forking working when run from inside the phar. The only solution I could find was to add a sleep right after a fork, which seemed to allow something to catch up and not end up with a corrupt phar (syntax errors where everywhere). The current phars are 500k compressed, 1.8M uncompressed. Stripping whitespace by itself makes them 960k, which will have to do.