Add sniff to disallow alternative PHP open tags.
This sniff checks  for the alternative PHP syntaxes - ASP tags and script tags - which have been removed in PHP7:
```php
<% echo 'content'; %>
<%= 'content'; %>
<script language="php">
echo 'content';
</script>
```
Ref: http://php.net/manual/en/language.basic-syntax.phptags.php

If PHPCS is run on PHP < 7, the script tags will be fixable, provided they are correct.
If PHPCS is run on PHP < 7 + asp_tags is on, the ASP tags will be fixable.

Fixes #1063

Similar to #1069, the `myconfig5.ini` and `myconfig57.ini` files and the adjustment to the travis script are needed for the fixer unit tests to pass as when asp_tags is set to off, the fixer won't fix it correctly and the tests will fail ;-)