Fix FunctionCallSignatureSniff
The `findNext` method doesn't return token at the `$end` position. In that case we want the `$end` token to be returned (the close bracket) so there should be `$closeBracket + 1`. At least I think so - it solved my problem. I debugged it couple times and this seems to be the valid and intended `$end`.