Play nice with sniff libraries with alternative unit test setups.
If there are `installed_paths` set up in the config and one of the standards in the `installed_paths` uses an alternative way to unit test - this applies in particular to the PHPCompatibility sniff library -, running the unit tests for any of the other standards using the normal PHPCS unit test set up will fail as fatal errors will be thrown for (sniff) files which cannot be found.

The solution consists of two parts based on various iterations while trying to solve this.

Part 1 is checking an environment variable which can be passed in through the phpunit.xml file which allows for indicating one or more standards to ignore when setting up the test suite.
This is the simplest way to prevent the issue.
As there won't be that many people working on several standards which have their own test suite, for those few people to add the environment variable to their phpunit.xml file for the projects which do use the PHPCS native test suite, should not be much of a problem.

Part 2 is not necessary, but might be a nice to have (=result of an earlier iteration).
This involves the catching of the file path names of the files which are causing issues and reporting them back as part of the phpunit results in a neat list without errors being thrown before starting to run the unit tests.