Generic/DuplicateClassName: fix property visibility
While working on the list of public properties which can be changed from rulesets, I came across this property for which it seemed unlikely that these were intended to be public.

This is a complex property which expects a multi-level array to be set based on the classes found by the sniff.