Generic / UnusedFunctionParams: Only allow for empty function bodies if function implements interface
The sniff currently allowed for unused parameters in function with empty bodies to allow for selective interface implementation.
The sniff, however, did not verify if the function was in a class which implemented an interface.
This has now been fixed.

Also minor efficiency tweak by not calling a function in a loop construct.

Includes unit test.