Generic / UnusedFunctionParams: Differentiate error codes for unused parameters before/after last used param.
(Callback) Functions may not use all parameters passed, but have to allow in their function signature for the ones they do use to be passed to them.

This commit introduces additional differentiation in the error codes used based on:
* whether the *only* param was not used (`Found`)
* whether a param was not used, but positioned *before* the last used param (`FoundBeforeLastUsed`)
* whether a param was not used and positioned *after* the last used param (`FoundAfterLastUsed`)

These codes will combine with the previously introduced `FoundInExtendedClass` and `FoundInImplementedInterface` error codes, resulting in the following error code list for this sniff:
* `Found`
* `FoundBeforeLastUsed`
* `FoundAfterLastUsed`
* `FoundInExtentedClass`
* `FoundInExtentedClassBeforeLastUsed`
* `FoundInExtentedClassAfterLastUsed`
* `FoundInImplementedInterface`
* `FoundInImplementedInterfaceBeforeLastUsed`
* `FoundInImplementedInterfaceAfterLastUsed`

Includes unit tests.