Tokens: Add new `$ooScopeTokens` property.
Add new `$ooScopeTokens` property representing all tokens which typically open class-like constructs.

This is particularly useful if a sniff needs to behave differently when it encounters a token within one of those scopes, like in the case of NamingConvention sniffs.

Applied the use of the property in a few places in existing sniffs.