:fire: Fix broken/incompatible `installed_paths` in PHPCS 3.x
If a "Standards" directory is called the same as one of the standards it contains and that standard is requested an _"ERROR: the ... coding standard is not installed."_ will be thrown and the sniffs will not run.

While I suppose this change in behaviour between PHPCS 3.x and PHPCS 2.x is intended to add support for "one standard" type of libraries, it breaks support for multi-standard libraries where one of the standards has the same name as the repository.

The change I'm proposing in this PR will maintain support for the PHPCS 2.x behaviour - thus not breaking existing registered standards - while at the same time supporting the new feature.