:fire: Fix broken sniff exclusions/restrictions in combination with custom standards in PHPCS 3.x
When running PHPCS with a custom standard, the command-line `--sniffs` and `--exclude` options no longer work.

This is caused by the sniffNames being prefixed with PHPCS's own `php_codesniffer\standards\` namespace which custom standard should not use in the first place.

This PR fixes this by only prefixing the sniffname with the PHPCS native namespace when it is a sniff from one of the build-in standards.