:fire: Fix PHPCS native autoloader & class recognition.
While working on making a custom standard compatible with PHPCS 3.x using `class_alias()`-es, a number of sniff files were getting misidentified by the autoloader causing PHPCS to choke.

This change fixes the issue I was encountering by using the the `$newClasses` diff in reverse.

While debugging this I notices copy/paste artefacts in the subsequent checks for new interfaces and traits. Both of which were - incorrectly - being compared against the `$classes` variable.
I've taken the liberty to rename the local variables related to this as well to make the distinction clearer.