Fix Exception on configuration-file errors
When accessing the configuration file to set values while it is not writ-
able, a Runtime exception was thrown but not converted into a DeepExit
exception.

Fix is to import the PHP Exception class in PHP_CodeSniffer\Config.

Additionally when writing to the file returns the error value false, the
method Config::setConfigData() returned false instead of throwing a
RuntimeException. Fix is to throw the exception as well.

Additionally when accessing all configuration data via include() while
the configuration file does not contain any definition of the variable
named $phpCodeSnifferConfig a warning was given for an unset variable. Fix
is to initialize the variable after inclusion if it was not yet set.

To implement the changes and to cover for regressions tests have been
added for the Config class.