DisallowSpaceIndent: Allow the sniff to fix space indents for multi-line `/* */` comments
Code sample:
```php
    /*
     * Some text
     */
```

The tokenizer tokenizes this code as:
```
T_WHITESPACE T_COMMENT
T_COMMENT
T_COMMENT
```

In other words, the whitespace at the start of subsequent comment lines is tokenized as part of the `T_COMMENT`.

In effect, this meant that these lines would not throw the appropriate error, nor be fixed by this sniff.

This commit fixes that.

Includes unit tests.