Squiz/InlineComment: Fix two more issues where errors are incorrectly not being thrown
If two subsequent lines contained comments with code between them, they would be seen as part of the same comment block, while in reality there are not.

This meant that errors for `NotCapital` and `InvalidEndChar` where not always being thrown when they should be.

Includes additional unit tests demonstrating the issue.