Generic/DisallowSpaceIndent: fix fixer bug when line only contains superfluous whitespace.
When the `Generic.WhiteSpace.DisallowSpaceIndent` sniff encountered spaces on an otherwise empty line, the fixer would replace the spaces with tabs, remove the new line character and in effect prefix the whitespace to the next line.

In other words, without the fix contains in this PR, this code snippet:
```php
// The next line has four spaces. This is a specific test case. Do not remove the spaces.

$bar = 2;
```
would be fixed as:
```php
// The next line has four spaces. This is a specific test case. Do not remove the spaces.
	$bar = 2;
```

Includes unit test demonstrating the issue.

The fix in this PR results in this sniff ignoring lines which only contain whitespace.
Superfluous whitespace can be trimmed from lines using this `Squiz.WhiteSpace.SuperfluousWhitespace` sniff after all.