Allow Generic/Sniffs/WhiteSpace/DisallowSpaceIndentSniff to support multiline alignment.
The Generic/Sniffs/WhiteSpace/DisallowSpaceIndentSniff erroneously tries to manipulate spaced alignment over multiple lines, where the number of spaces exceeds `$tabWidth`.

This pull request fixes this issue by trying to be smart about remembering the indentation of the previous line and ignoring the spaces that are supposed to be for alignment only. They need to meet the following criteria:

* Same tabbed indentation than on previous line.
* No tabs after spaces.

Fixes #1735