Fixed bug #1462 : Error processing cyrillic strings in Tokenizer
The error from iconv_strlen when a string contains invalid chars (based on the encoding) was no longer being muted due to the new error handler in the Runner class. This commit replaces the mute operator with an error_reporting change to properly mute that error again and allow files to be checked even with mixed encoding.