Add link to download the phar files in the README
There was a link pointing to the Phar documentation, which I confused with the link to download the actual phars. I think this link is confusing and useless. Instead it makes much more sense to have a link to the files to download directly (I spent several minutes trying to find where to download those files).