PSR2/SwitchDeclaration: bug fix - ignore PHPCS annotation tokens
Treat the new `// phpcs:` comments in the same way as "ordinary" `T_COMMENT` tokens.

Includes adjusted unit test.

This fixes a bug where when a `// phpcs:ignore` comment is at the end of a line, it would be incorrectly "fixed" by moving it down a line, making the comment apply to a different line altogether.