Generic/LineLength: ignore new PHPCS annotations for the metrics
Ignore the new `// phpcs:` annotation comments completely when they are on a line by themselves.
Do not record metrics about lines only containing these type of comments.

Errors would not be thrown anyway as lines containing these comments are ignored when throwing errors. Still, this does save the sniff from executing the rest of the logic when its not necessary.

Includes unit test.