Tokenizer/ErrorSuppression: `phpcs:set` only ignore line when PHPCS annotation is complete line
This fixes a situation where a `phpcs:set` could be used at the end of a line and the whole line was always being ignored.

```php
do_something(); //phpcs:set Standard.Category.Sniff propertyName value
```

I wasn't sure where to add a unit test for this as there did not seem to be any for `phpcs:set` in the `errorSuppression.Test.php` file.