Squiz/BlockComment: fix issue with PHPCS annotations causing the comment closer to be misidentified.
Whether something was identified as a comment line was based on the token code being the same.
This did not allow for PHPCS annotations to be included in block comments as those have a different token code.

This resulted in comments being "cut short", i.e. the line above a PHPCS annotation being seen as the last line in a comment and subsequent comment lines being ignored, resulting in incorrect `Comment closer must be on a new line` and `Empty line required after block comment` notices.

Fixes 1918