Squiz/ControlSignature: improve handling of do/while structures
Up to now, `do/while` structures would be examined based on the `do` token, while all other multi-keyword control structures were examined based on the keyword *after* the closing brace.

As the fix in the previous commit depends on `$stackPtr` pointing to the keyword *after* the closing brace, the handling of `do/while` structures has been made consistent with how the other multi-keyword structures are handled.