Generic/DocComment: improve handling of PHPCS annotations
When PHPCS is run with `--ignore-annotations`, the `Generic.Commenting.DocComment` sniff would create "noise" complaining about PHPCS annotation comment lines not starting with a capital letter.
It would also - incorrectly - *not* complain about a missing short description if the phpcs annotation is the first thing in a docblock.

This change fixes both these issues.

Includes unit tests in a separate file to allow for the config settings for that file to be different from the main test file(s).