Squiz/BlockComment: improve handling of PHPCS annotations
When PHPCS is run with `--ignore-annotations`, the `Squiz.Commenting.BlockComment` sniff would create a lot of "noise" complaining about PHPCS annotation comment lines not starting with a capital letter or ending in a full stop.

This change prevents those errors being thrown when a PHPCS annotation is encountered.

Includes unit tests in a separate file to allow for the config settings for that file to be different from the main test file(s).