ConstantDeclaration: simplify the unit tests
The sniff unit tests should test the functioning of the sniff which can be done perfectly well using the PHP version as used to run the sniff.

Looks like what you are/were trying to do now was basically testing whether the `Config` class works as expected from within a sniff and while I agree, it may be a good idea to add unit tests for the `Config` class, this is not the task of a sniff to do.

This PR simplifies the unit test code to defer to the PHP version used to run the sniff to test it.