Squiz/OperatorSpacing: fix fixer conflict
[Fixer conflicts series PR]

Fixer conflict can be reproduced by running the below command against `master`:
`phpcbf -p -s ./src/Standards/Squiz/Tests/Arrays/ArrayDeclarationUnitTest.1.inc --standard=Squiz`

When there is a comment after an operator, the `Squiz.Commenting.PostStatementComment` sniff and the `Squiz.WhiteSpace.OperatorSpacing` sniffs would conflict, with the `PostStatementComment` sniff demanding a new line and the `OperatorSpacing` sniff demanding exactly one space.

Along the same lines, if the next content is a PHPCS annotation on the next line, moving the annotation to be on the same line as the operator could change its meaning.

I've chosen to fix this by throwing the error, but blocking auto-fixing when the next thing after the operator is a comment/annotation and the whitespace found is a new line.

Includes unit test.