Generic/UnnecessaryFinalModifier: test valid code
The unit tests for this sniff tested properties with the `final` keyword, however, that is not a valid syntax. The `final` keyword can only be applied to classes and methods.

These changes, make the unit tests more realistic.

The sniff itself reports on finding the `final` keyword within a `final` class independently of the context in which the keyword is used, so did not need any changes.