Generic/UnnecessaryStringConcat: fix test which wasn't testing anything
The code `$x = 'My '.1234;` is a parse error as the `.` would tokenize together with the number to a `T_DNUMBER` and not to a `T_STRING_CONCAT`, which means nothing was actually being tested. By adding the space, the `.` is now correctly tokenized which activates the test that this _doesn't_ give an error properly.