Squiz/FunctionSpacing: don't remove indentation before start of function
When too many blank lines are found before a function, the `Before/BeforeFirst` fixer would unnecessarily also remove the indentation of the `$nextContent` line, meaning that this sniff would **always** need to be accompanied by one or more sniffs to fix the indentation again.

As this sniff is about blank lines between functions, not about indentation, I consider this a bug.

This minor change fixes it.

Includes unit tests demonstrating the issue.