Fixer: very minor efficiency fix
Since PHP 5.3.6, `debug_backtrace()` can be called with the `DEBUG_BACKTRACE_IGNORE_ARGS` option.
By default, `debug_backtrace()` when called from within a function will provide access to the parameters passed to the function.
Using the `DEBUG_BACKTRACE_IGNORE_ARGS` option, it won't and as the `args` index from the resulting backtrace arrays is not used in the `Fixer` anyway, using that option should lower the memory usage a tiny bit.