Better Fix for issues with closures.
some functions are not like  `$this->attachRule(` considered a condition and therefore doesn't count towards the number of levels. Instead it's considered a nested_parenthesis (whatever that means) and the same for the `usort` function in the other unit test we had. We blindly therefore added one level to the count; but instead we should actually check for the nested_parenthesis and count how many of them there are 
This fix was created by @wilsonge