Tokenizer: fix binary cast recognition with inside parentheses whitespace
As reported in my last comment on the original issue 1574, the fix to the tokenizer does not take whitespace inside of the binary cast parentheses into account.

This PR fixes this.

It also adds a unit test which will guard against regressions to the `Squiz.WhiteSpace.CastSpacing` sniff.

As that sniff did not have a `fixed` file, while it _does_ in actual fact contain a fixer, the `.fixed` file has also been added.