Squiz.CSS.SemicolonSpacing: add fixed file and various improvements
The sniff contains fixers, but didn't have a `fixed` file.

While I was at it, I've reviewed the sniff and applied any improvements I could think of:
* The sniff will now bow out during live coding.
* The sniff did not take multiple CSS declarations on a single line into account.
* The sniff did not take multi-line CSS declarations into account.
    The error messages will now be thrown on the last line of the CSS declaration, not on the line containing the `T_STYLE` token.
* The sniff did not take multi-line whitespace between the end of a declaration and the semi-colon into account.
    - The error message has been adjusted to provide better information in that case.
    - The fixer has been adjusted to now fix this in one go.
* The sniff did not take comments or PHPCS annotations into account.
    - The sniff has been adjusted to not get confused over this anymore.
    - The fixer has been adjusted to handle comments correctly.

Includes additional unit tests for all fixes.

N.B.: Working on this left me wondering why the `NotAtEnd` (missing semi-colon) error is not auto-fixable. I couldn't find anything in the file history related to this and with the code as it is in this PR, adding a fixer for the missing semi-colon would actually be really easy, so let me know if that would be desired.