Squiz/ForLoopDeclaration: fix fixer conflict
Fix a fixer conflict within the sniff.

If `$requiredSpacingAfterOpen/BeforeClose` would be set to anything but `0` and the first or the last clause of the `for()` would be empty, the sniff had a fixer conflict with itself, as the `SpacingAfterOpen/BeforeClose` error would demand whitespace, while the `SpacingBeforeFirst/AfterSecondNoThird` would try to remove it.

Fixed by skipping the semicolon spacing checks when an empty first/last expression is detected and leaving the handling of those situations up to the `SpaceAfterOpen/BeforeClose` logic.

This fix effectively removes the `SpacingAfterSecondNoThird` error code.