Squiz/ForLoopDeclaration: prevent creating parse error with `SpacingBeforeClose` fixer
When a `//`-style comment would be at the end of a multi-line `for` declaration, the `SpacingBeforeClose` fixers could inadvertently create a parse error, by moving the `for` close parenthesis into the comment.

Fixed by checking whether the previous non-whitespace token is on another line and a comment and making the errors non-fixable in that case.