Travis: add check to make sure that sniffs are "feature complete"
This PR adds a custom script which will check whether each and every sniff is accompanied by a documentation XML file (warning) as well as unit test files (error).

Notes:
* To only show errors, the script can be run in `quiet` mode by passing the `-q` flag on the command line.
* The script has been set up in such a way that it is not only suitable for use by PHPCS itself, but also for use by external standards which use the PHPCS native unit test suite.
    For external standards to use the tool, run it from the root of the external standards repo like so:
    `php -f "path/to/PHP_CodeSniffer/scripts/check-sniff-completeness.php"`
* The list of sniffs to check completeness for is determined by getting a list of all files in `Sniffs` directories, where the files have a name ending with `Sniff.php` and not prefixed with `Abstract`.
* The "has tests" check verifies that a `UnitTest.php` file is present in the standard's `Tests` directory and _at least_ one test case file.
* The "has docs" check verifies that a `Standard.xml` file is present in the standard's `Docs` directory.

This check only needs to be run on one build as the results won't change between PHP versions.
For that reason, I've elected to add it to the fastest PHP version run to balance out the runs some more.

As quite a large number of sniffs at this moment do not have documentation files, `quiet` mode is currently enabled for the check in Travis.
If/when the missing documentation files have been added, the `-q` flag should be removed and having unit tests as well as documentation should be enforced for all sniffs.